Refactor selectors to enhance state management with improved performance and memory handling
